+++ source/api/Meat.Application/Sucursales/DeleteSucursal/DeleteSucursalHandler.cs
@@ -25,12 +25,21 @@ namespace Meat.Application.Sucursales.DeleteSucursal
 
             var sucursal = await this.context.Sucursales
                 .Include(x => x.Puestos)
-                .FirstOrDefaultAsync(x => x.Id == request.Id);
+                .Include(x => x.Empresa)
+                .FirstOrDefaultAsync(x => x.Id == request.Id && x.Empresa.CodigoEmpresa == request.CodigoEmpresa);
 
             if (sucursal == null)
             {
                 throw new ValidationException("La sucursal no existe");
             }
+
+            var tieneEstablecimientos = await this.context.Establecimientos.AnyAsync(e => e.SucursalId == request.Id, cancellationToken);
+            if (tieneEstablecimientos)
+                throw new ValidationException("No se puede eliminar la sucursal porque tiene establecimientos asignados.");
+
+            var tieneUsuarios = await this.context.UsuariosSucursales.AnyAsync(us => us.SucursalId == request.Id, cancellationToken);
+            if (tieneUsuarios)
+                throw new ValidationException("No se puede eliminar la sucursal porque tiene usuarios asignados.");
 
             this.context.Sucursales.Remove(sucursal);
